+++ tools/api-markdown-documenter/src/renderers/html-renderer/test/RenderDocument.test.ts
@@ -3,37 +3,57 @@
  * Licensed under the MIT License.
  */
 
+import type { ApiItem } from "@microsoft/api-extractor-model";
 import { expect } from "chai";
 
-import {
-	DocumentNode,
-	HeadingNode,
-	ParagraphNode,
-	PlainTextNode,
-	SectionNode,
-	SpanNode,
-} from "../../../documentation-domain/index.js";
+import type { ApiDocument } from "../../../ApiDocument.js";
+import { HeadingNode, SectionNode } from "../../../documentation-domain/index.js";
 import { renderDocument } from "../Render.js";
 
 describe("Document HTML rendering tests", () => {
 	it("Renders a simple document", () => {
-		const document = new DocumentNode({
-			children: [
+		const document: ApiDocument = {
+			apiItem: {} as unknown as ApiItem, // Mock ApiItem for testing
+			contents: [
 				new SectionNode(
 					[
-						new ParagraphNode([
-							new PlainTextNode("This is a sample document. "),
-							new PlainTextNode("It has very basic content.\t"),
-						]),
+						{
+							type: "paragraph",
+							children: [
+								{
+									type: "text",
+									value: "This is a sample document. ",
+								},
+								{
+									type: "text",
+									value: "It has very basic content.\t",
+								},
+							],
+						},
 						new SectionNode(
 							[
-								new ParagraphNode([
-									new PlainTextNode("This is text inside of a paragraph. "),
-									new PlainTextNode("It is also inside of a hierarchical section node. "),
-									SpanNode.createFromPlainText("That's real neat-o.", {
-										italic: true,
-									}),
-								]),
+								{
+									type: "paragraph",
+									children: [
+										{
+											type: "text",
+											value: "This is text inside of a paragraph. ",
+										},
+										{
+											type: "text",
+											value: "It is also inside of a hierarchical section node. ",
+										},
+										{
+											type: "emphasis",
+											children: [
+												{
+													type: "text",
+													value: "That's real neat-o.",
+												},
+											],
+										},
+									],
+								},
 							],
 							new HeadingNode("Section Heading"),
 						),
@@ -42,7 +62,7 @@ describe("Document HTML rendering tests", () => {
 				),
 			],
 			documentPath: "./test",
-		});
+		};
 
 		const expected = [
 			"<!doctype html>",
@@ -56,7 +76,7 @@ describe("Document HTML rendering tests", () => {
 			"      <p>This is a sample document. It has very basic content.</p>",
 			"      <section>",
 			"        <h2>Section Heading</h2>",
-			"        <p>This is text inside of a paragraph. It is also inside of a hierarchical section node. <i>That's real neat-o.</i></p>",
+			"        <p>This is text inside of a paragraph. It is also inside of a hierarchical section node. <em>That's real neat-o.</em></p>",
 			"      </section>",
 			"    </section>",
 			"  </body>",