+++ packages/drivers/odsp-urlResolver/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @fluidframework/odsp-urlresolver
 
+## 2.52.0
+
+Dependency updates only.
+
 ## 2.51.0
 
 Dependency updates only.