+++ packages/test/mocha-test-setup/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @fluid-internal/mocha-test-setup
 
+## 2.52.0
+
+Dependency updates only.
+
 ## 2.51.0
 
 Dependency updates only.