+++ packages/utils/odsp-doclib-utils/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @fluidframework/odsp-doclib-utils
 
+## 2.52.0
+
+Dependency updates only.
+
 ## 2.51.0
 
 Dependency updates only.